Rename ActionsStack controller
